feat(16216): add GET Listeners hook - rename folder for frontend services
