+++ hivemq-edge/src/frontend/src/api/hooks/useFrontendServices/useGetConfiguration.tsx
